Split pamwire by responsibility: stanzas, grammar, transform, files, report
pamwire.rs had grown to ~4800 lines holding five unrelated jobs: the exact
bytes written into an auth stack, reading a stack, rewriting one, writing it
to disk safely, and reporting on it. Each is separately testable and none
needed to see the others' internals.

  stanzas.rs    the PAM lines we write, and the module names we look for
  grammar.rs    predicates over a stack line; where the face block anchors
  transform.rs  &str -> (String, bool) rewrites, no I/O
  files.rs      staged writes, backups, restore, the PAM lock
  report.rs     login status, surface facts, the keyring-handoff warning
  pamwire.rs    which surfaces to wire, and the enable/disable/reconcile flow

Deliberately NOT split by modality. Face and fingerprint lines share the same
greeter files -- a GDM box gets the face `unseal` line and the fingerprint
keyring line in one /etc/pam.d/gdm-password, in a required order -- so a
face.rs/fingerprint.rs split would put a single ordering invariant under two
owners, which is the shape of bug the preceding commits spent their time
fixing. The device layer is already separated where it belongs, in the
irlume-fingerprint crate.

Pure movement. Bodies are unchanged; the only edits are visibility widened to
pub(super) for items now referenced across the boundary, module declarations,
and an explicit re-export list for the pub(crate) surface the rest of the CLI
reaches as pamwire::*. A glob `use` binds names privately, so that surface is
now stated in one place instead of inherited.

The test module is deliberately left in pamwire.rs and untouched, so the same
367 assertions run against the moved code and the diff stays reviewable as
movement rather than rewriting.

Verified byte-identical: the eight upstream stacks these commits added as
fixtures (plasmalogin fedora/arch/debian/suse, gdm-password redhat/arch/
renamed-substack, gdm-fingerprint) were generated from the commit before this
one and from this one, and diff clean.
