+++ crates/irlume-cli/src/pamwire.rs
@@ -22,73 +22,33 @@ use irlume_common::platform::{distro_family, DistroFamily};
 use std::path::{Path, PathBuf};
 use std::process::{Command, ExitCode};
 
-const MODULE: &str = "pam_irlume.so";
-pub(crate) const BACKUP: &str = ".pre-irlume";
-const CREATED_PREFIX: &str = "# irlume: created from ";
-/// The one sentence that explains the on-demand trigger; shared so the status
-/// line, the plan line, and docs/SETUP.md's mirror never drift apart.
-const ONDEMAND_HINT: &str = "leave the password empty and press Enter to use your face";
+// Horizontal split by responsibility, innermost first: the bytes we write
+// (`stanzas`), reading a stack (`grammar`), rewriting one (`transform`). All
+// three are pure — no filesystem, no policy — which is what keeps this module's
+// file handling and the wiring decisions testable apart from each other.
+//
+// Deliberately NOT split by modality: face and fingerprint lines share the same
+// greeter files (a GDM box gets the face `unseal` line and the fingerprint
+// keyring line in one `/etc/pam.d/gdm-password`, in a required order), so a
+// face/fingerprint split would put one ordering invariant under two owners.
+mod files;
+mod grammar;
+mod report;
+mod stanzas;
+mod transform;
 
-// Greeter block for a non-`@include` (Fedora `substack`) stack: a `success=1`
-// jump over the password substack, plus the `PERMIT_LANDING` it lands on.
-/// Jump-style face line for a submit-driven greeter we have NOT validated for the
-/// on-demand probe (old GDM, unknown DM). `facefirst` is mandatory here: without
-/// it the module runs the ACTIVE probe (`pam_get_authtok`), and a greeter that
-/// blocks that probe until the user types means face never fires at all.
-const GREETER_UNSEAL_FACEFIRST_JUMP: &str =
-    "auth       [success=1 default=ignore]   pam_irlume.so unseal facefirst";
-/// The greeter/locker face line for any INCLUDE layout: Debian/Ubuntu
-/// `@include common-auth` and Arch `auth include system-login` alike (a
-/// `success=N` jump can't skip an include expansion, so this can't be the jump
-/// form). Always `sufficient`: the same control works for EVERY DM's locker
-/// (GDM and cosmic alike short-circuit on a warm unlock). Cold-login keyring
-/// unlock is handled by the module's `kr` arg, NOT the control: on a cold login
-/// the module returns IGNORE (having set the token), so `sufficient` continues to
-/// pam_unix + pam_gnome_keyring; a warm lock returns SUCCESS and short-circuits.
-/// `mode` is `facefirst` (GDM scan-immediately) or `ondemand` (empty-Enter). `kr`
-/// adds the keyring-continue arg: true for greeters (cold login unlocks the
-/// keyring), false for a separate warm lock service (keyring already open).
-fn include_greeter_line(mode: &str, kr: bool) -> String {
-    let kr_arg = if kr { " kr" } else { "" };
-    format!("auth       sufficient   pam_irlume.so unseal {mode}{kr_arg}")
-}
-/// Jump-style variant for a non-`@include` (e.g. Fedora) COSMIC stack.
-const GREETER_UNSEAL_COSMIC_JUMP: &str =
-    "auth       [success=1 default=ignore]   pam_irlume.so unseal ondemand";
-// Tagged so unwire strips OUR landing but never a foreign pam_permit.so the
-// stack legitimately carries (the trailing `#…` is a PAM comment, ignored).
-const PERMIT_LANDING: &str =
-    "auth       optional                     pam_permit.so   # irlume-landing";
-const RESEAL_AUTH: &str = "auth       optional                     pam_irlume.so reseal";
-/// Post-auth login-keyring unlock for the FINGERPRINT path: runs after a trusted
-/// factor succeeded; if no password is present (fingerprint login) it unseals
-/// the TPM-sealed password and sets PAM_AUTHTOK so pam_gnome_keyring opens the
-/// wallet. No-op when the keyring isn't armed or a password is already set.
-const KEYRING_UNSEAL: &str = "auth       optional                     pam_irlume.so keyring";
-/// Tag on the gnome-keyring lines irlume ADDS to a fingerprint stack, so unwiring
-/// removes exactly ours and never a keyring line the distro shipped. Linux-PAM
-/// strips a trailing `#` comment before tokenizing (verified against
-/// `pam_exec.so`: an argument survives, a trailing comment does not), so this is
-/// invisible to the module — which matters here because gnome-keyring's
-/// `parse_args` syslogs a warning for every option it does not recognize.
-const KEYRING_TAG: &str = "# irlume-keyring";
-/// GDM's `gdm-fingerprint` stack carries NO keyring module at all (verified on
-/// upstream `data/pam-redhat/gdm-fingerprint.pam` through GDM 50.0), so the
-/// `KEYRING_UNSEAL` line above would set a token nothing reads. These are the
-/// two halves gnome-keyring needs, added only when the stack has no consumer of
-/// its own. The leading `-` is PAM's "do not complain if the module is missing",
-/// so a machine without gnome-keyring installed is unaffected.
-const FP_GKR_AUTH: &str =
-    "-auth      optional                     pam_gnome_keyring.so   # irlume-keyring";
-const FP_GKR_SESSION: &str =
-    "-session   optional                     pam_gnome_keyring.so auto_start   # irlume-keyring";
-const RESEAL_SESSION: &str = "session    optional                     pam_irlume.so reseal";
-/// The plain verify stanza, shared by `sudo` and polkit prompts (Bitwarden vault
-/// unlock, pkexec, systemd unit control): no `unseal` (the daemon refuses
-/// credential release for both classes anyway) and no mode arg (each surface runs
-/// the PAM conversation as soon as it prompts, which IS the face-first trigger;
-/// the daemon adds the forced consent gesture on top).
-const VERIFY_STANZA: &str = "auth       sufficient                   pam_irlume.so";
+use files::*;
+use grammar::*;
+use report::{report_keyring_handoff, status};
+use stanzas::*;
+use transform::*;
+
+// Re-exported for the rest of the CLI, which reaches these as `pamwire::…`.
+// A glob `use` binds names privately, so the public surface is listed here
+// rather than inherited — which also keeps that surface visible in one place.
+pub(crate) use files::{is_managed_path, lock_pam, restore_surface};
+pub(crate) use report::{login_manager_fact, status_report, surface_facts};
+pub(crate) use stanzas::BACKUP;
 
 /// A PAM service to wire. `vendor=Some` → materialize an /etc override from the
 /// vendor copy; `vendor=None` → back up and edit the real /etc file.
@@ -178,33 +138,6 @@ pub fn run(action: Option<&str>, args: &[String]) -> ExitCode {
             ExitCode::from(2)
         }
     }
-}
-
-/// Structured wiring status for the TUI: `(label, present, wired)` per service
-/// plus a trailing SELinux row. Mirrors what `status()` prints.
-pub(crate) fn status_report() -> Vec<(String, bool, bool)> {
-    let mut out = Vec::new();
-    for s in GREETERS
-        .iter()
-        .chain(FP_GREETERS.iter())
-        .chain(std::iter::once(&LOCKSCREEN))
-    {
-        match service_present(s) {
-            Some(p) => out.push((label_of(s.etc), true, file_has_module(&p))),
-            None => out.push((label_of(s.etc), false, false)),
-        }
-    }
-    let sudo = Path::new(SUDO);
-    out.push((
-        "sudo".into(),
-        sudo.exists(),
-        sudo.exists() && file_has_module(sudo),
-    ));
-    match service_present(&POLKIT) {
-        Some(p) => out.push(("polkit (apps)".into(), true, file_has_module(&p))),
-        None => out.push(("polkit (apps)".into(), false, false)),
-    }
-    out
 }
 
 /// True when any greeter or the lock screen carries the irlume wiring; the
@@ -490,11 +423,6 @@ pub(crate) fn polkit_wired() -> Option<bool> {
 /// name the machine output can publish as a stable id.
 fn service_name(etc: &str) -> &str {
     etc.rsplit('/').next().unwrap_or(etc)
-}
-
-/// Short label from an /etc/pam.d path (e.g. "/etc/pam.d/gdm-password" → "gdm-password").
-fn label_of(etc: &str) -> String {
-    service_name(etc).to_string()
 }
 
 /// The active login manager, from the `display-manager.service` symlink
@@ -789,82 +717,6 @@ const ROLE_LOCK: &str = "lock-screen";
 const ROLE_SUDO: &str = "sudo";
 const ROLE_POLKIT: &str = "polkit";
 
-/// One PAM surface irlume knows how to wire, as facts rather than a rendered
-/// row. The human report prints these and `login status --json` serializes
-/// them, so a single pass feeds both and the two cannot disagree about what is
-/// wired; they can only differ in how they word it.
-pub(crate) struct SurfaceFact {
-    /// The PAM service name (`plasmalogin`, `kde`, `sudo`, …). Stable public id.
-    pub(crate) id: &'static str,
-    pub(crate) role: &'static str,
-    /// The /etc path, for the human report only. Machine output publishes the
-    /// service name instead, in keeping with the no-paths rule.
-    pub(crate) path: &'static str,
-    /// Whether this service exists here at all: a real /etc file, or a vendor
-    /// copy an /etc override would be materialized from.
-    pub(crate) present: bool,
-    pub(crate) wired: bool,
-    /// How face fires here when wired: `face-first`, `on-demand`, `keyring`
-    /// (the fingerprint keyring-unlock line, which is not a face factor) or
-    /// `verify` (the plain sudo/polkit stanza). `None` when not wired.
-    pub(crate) mode: Option<&'static str>,
-}
-
-fn surface_fact(
-    etc: &'static str,
-    vendor: Option<&'static str>,
-    role: &'static str,
-) -> SurfaceFact {
-    let path = service_present(&Svc { etc, vendor });
-    let content = path
-        .as_ref()
-        .and_then(|p| std::fs::read_to_string(p).ok())
-        .unwrap_or_default();
-    let wired = content_has_module(&content);
-    // Name HOW face fires on this service: on-demand (the consent model) is
-    // invisible in the PAM file to a reader, and a keyring-only line is the
-    // fingerprint unlock rather than face at all.
-    let mode = if !wired {
-        None
-    } else if role == ROLE_SUDO || role == ROLE_POLKIT {
-        Some("verify")
-    } else if !content.contains("unseal") {
-        Some("keyring")
-    } else if content.contains("ondemand") {
-        Some("on-demand")
-    } else {
-        Some("face-first")
-    };
-    SurfaceFact {
-        id: service_name(etc),
-        role,
-        path: etc,
-        present: path.is_some(),
-        wired,
-        mode,
-    }
-}
-
-/// Every surface irlume can wire, present or not, in the order the human report
-/// prints them. Absent services are kept in the list with `present: false`: a
-/// consumer must be able to read an id it knows and cannot find as "this engine
-/// does not wire that service" rather than as "not wired here".
-pub(crate) fn surface_facts() -> Vec<SurfaceFact> {
-    let mut out: Vec<SurfaceFact> = GREETERS
-        .iter()
-        .map(|s| surface_fact(s.etc, s.vendor, ROLE_LOGIN))
-        .chain(
-            FP_GREETERS
-                .iter()
-                .map(|s| surface_fact(s.etc, s.vendor, ROLE_LOGIN_FP)),
-        )
-        .collect();
-    out.push(surface_fact(LOCKSCREEN.etc, LOCKSCREEN.vendor, ROLE_LOCK));
-    out.push(surface_fact(SUDO, None, ROLE_SUDO));
-    out.push(surface_fact(POLKIT.etc, POLKIT.vendor, ROLE_POLKIT));
-    out
-}
-
 /// Which factors this machine's hardware and configured method call for.
 ///
 /// Extracted so the human `login enable` report and the machine plan derive
@@ -905,29 +757,6 @@ pub(crate) struct PlannedSurface {
     /// carrying that id would overwrite a stack the consumer was never shown.
     /// The digest is what makes the id describe a state rather than an intent.
     pub(crate) state: String,
-}
-
-/// A surface's current content digest, or `ABSENT`, or `unreadable`.
-///
-/// Three distinct answers on purpose. Folding "cannot read" into either of the
-/// others would let a plan id stay stable across a state it could not actually
-/// observe.
-pub(crate) fn surface_state(path: &Path) -> String {
-    // The backup as well as the live file. Wiring rebuilds from `.pre-irlume`
-    // when one exists, so the content an apply produces depends on it: a backup
-    // that changed between the plan and the apply changes the outcome while the
-    // live file, and therefore the plan id, stayed identical. The consumer would
-    // be shown one result and the machine would get another.
-    let bak = PathBuf::from(format!("{}{BACKUP}", path.display()));
-    format!("{} {}", surface_digest(path), surface_digest(&bak))
-}
-
-pub(crate) fn surface_digest(path: &Path) -> String {
-    match crate::logintx::file_sha256(path) {
-        Ok(Some(digest)) => digest,
-        Ok(None) => crate::logintx::ABSENT.to_string(),
-        Err(_) => "unreadable".to_string(),
-    }
 }
 
 /// What `login enable`/`login disable` would change, computed without writing.
@@ -1259,182 +1088,6 @@ pub(crate) fn apply(
         },
     );
     out
-}
-
-/// Whether irlume manages this path at all.
-///
-/// A transaction record names the paths a rollback will write, and nothing
-/// previously checked that those were paths irlume had any business touching. A
-/// record naming /etc/shadow with a correct digest rewrote it: verified, not
-/// theorised. Only root can plant a record, and root can already write that
-/// file, so it was not an escalation, but it made `login rollback` a
-/// general-purpose write-anywhere-as-root primitive whose only gate was a
-/// directory mode. Any future way to plant a record would then be total.
-///
-/// So the paths are checked against the surfaces irlume wires, plus their
-/// `.pre-irlume` sidecars, and nothing else is restorable.
-pub(crate) fn is_managed_path(path: &str) -> bool {
-    let bare = path.strip_suffix(BACKUP).unwrap_or(path);
-    // Built from the same lists the wiring uses, so a surface added there is
-    // restorable without anyone remembering to update a second list.
-    GREETERS
-        .iter()
-        .chain(FP_GREETERS.iter())
-        .map(|s| s.etc)
-        .chain([LOCKSCREEN.etc, POLKIT.etc, SUDO])
-        .any(|managed| managed == bare)
-}
-
-/// Put one surface back to the content recorded before a transaction.
-///
-/// Reuses the same atomic write the wiring path uses, so a restore lands the
-/// way every other PAM write here does. The caller must have checked
-/// `unchanged_since_apply` first; this does the write, not the decision.
-///
-/// `None` content means the file did not exist before, so it is removed rather
-/// than written empty: an empty PAM file is not the same as an absent one, and
-/// leaving one behind would shadow a vendor copy.
-pub(crate) fn restore_surface(
-    path: &Path,
-    before: Option<&str>,
-    metadata: Option<(u32, u32, u32)>,
-) -> Result<(), String> {
-    match before {
-        Some(content) => {
-            // The recorded mode and owner go on before the rename, not after.
-            // Applying them afterwards published a PAM stack that was briefly
-            // whatever the root umask produced, and on this path in particular
-            // the file may have been REMOVED by apply, so there was nothing to
-            // copy attributes from and the default was all it ever got.
-            //
-            // `None` keeps the old behaviour for records written before those
-            // fields existed: the replacing file inherits the current one's
-            // attributes rather than a guess.
-            let attrs = metadata.or_else(|| {
-                std::fs::symlink_metadata(path).ok().as_ref().map(|m| {
-                    use std::os::unix::fs::MetadataExt as _;
-                    use std::os::unix::fs::PermissionsExt as _;
-                    (m.permissions().mode() & 0o7777, m.uid(), m.gid())
-                })
-            });
-            write_atomic_inner(path, content, attrs)
-        }
-        None => {
-            // The same refusal the replacing branch gets. Removing was a direct
-            // `remove_file`, so a path recorded as previously absent that is now
-            // a symlink was unlinked despite the claim that every write path
-            // refuses one, and a multiply-linked file lost a name irlume cannot
-            // put back.
-            inspect_target(path)?;
-            match std::fs::remove_file(path) {
-                Ok(()) => {}
-                Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-                Err(error) => return Err(format!("remove {}: {error}", path.display())),
-            }
-            // A deletion is a directory change like any other. Without this the
-            // unlink could be lost to a power cut while the durable progress
-            // note said the surface was done, so a resume would skip a file that
-            // is still there.
-            fsync_dir(path.parent().unwrap_or_else(|| Path::new(".")))
-        }
-    }
-}
-
-/// The active login manager and the PAM services it consults.
-pub(crate) struct LoginManagerFact {
-    /// `None` when no login manager could be found at all: a headless host. A
-    /// greeter that registers no `display-manager.service` is still found when
-    /// it is one of [`WANTS_ONLY_DMS`]. Not the same as "no face login".
-    pub(crate) name: Option<String>,
-    /// Whether irlume can wire face login here. False covers both a login
-    /// manager it has no mapping for and one whose mapped service it has no
-    /// recipe for; either way `login enable` cannot target it.
-    pub(crate) recognized: bool,
-    /// The greeter service, plus the separate fingerprint service on the login
-    /// managers that have one (GDM). Empty when unrecognized.
-    pub(crate) services: Vec<&'static str>,
-}
-
-pub(crate) fn login_manager_fact() -> LoginManagerFact {
-    let Some(dm) = active_display_manager() else {
-        return LoginManagerFact {
-            name: None,
-            recognized: false,
-            services: Vec::new(),
-        };
-    };
-    let (greeter, fp) = dm_pam_services(&dm);
-    // Named is not the same as wirable, so a service irlume cannot write must
-    // not be published as one it consults.
-    let recognized = dm_wirable(&dm);
-    LoginManagerFact {
-        name: Some(dm),
-        recognized,
-        services: if recognized {
-            std::iter::once(greeter).chain(fp).collect()
-        } else {
-            Vec::new()
-        },
-    }
-}
-
-fn status() -> ExitCode {
-    println!("[login] wiring status (face auth in PAM):");
-    if let Some(dm) = active_display_manager() {
-        let (greeter, fp) = dm_pam_services(&dm);
-        match fp {
-            Some(fp) => println!("  active login manager: {dm}  (uses {greeter} + {fp})"),
-            None => println!("  active login manager: {dm}  (uses {greeter})"),
-        }
-    }
-    let mut any = false;
-    let mut any_ondemand = false;
-    for f in surface_facts() {
-        if !f.present {
-            continue;
-        }
-        let label = match (f.role, f.mode) {
-            (ROLE_SUDO, Some(_)) => "● wired (sudo)",
-            (ROLE_SUDO, None) => "○ not wired (sudo)",
-            (ROLE_POLKIT, Some(_)) => "● wired (polkit app prompts)",
-            (ROLE_POLKIT, None) => "○ not wired (polkit app prompts)",
-            (_, None) => "○ not wired",
-            (_, Some("on-demand")) => {
-                any_ondemand = true;
-                "● wired (face on-demand)"
-            }
-            (_, Some("face-first")) => "● wired (face-first)",
-            // keyring-only line
-            (_, Some(_)) => "● wired",
-        };
-        // face-sudo alone does not make the login screen work, so it does not
-        // silence the "enable with" hint below.
-        any |= f.wired && f.role != ROLE_SUDO && f.role != ROLE_POLKIT;
-        println!("  {:<34} {}", f.path, label);
-    }
-    if any_ondemand {
-        println!("  on-demand: {ONDEMAND_HINT}");
-    }
-    // A greeter can be correctly wired and still leave the wallet locked, so
-    // this is reported next to the wiring rather than left to `doctor`: it is
-    // the difference between "face logs me in" and "face logs me in AND my
-    // wallet is open", which is what the keyring path exists for.
-    report_keyring_handoff();
-    println!(
-        "[login] SELinux module: {}",
-        match selinux_loaded() {
-            Some(true) => "loaded",
-            Some(false) => "not loaded",
-            None => "unknown (run as root to check)",
-        }
-    );
-    if !any {
-        println!(
-            "  → enable with:  sudo irlume login enable --apply   (add --with-sudo for face-sudo, \
-             --with-polkit for app prompts like Bitwarden)"
-        );
-    }
-    ExitCode::SUCCESS
 }
 
 fn act(enable: bool, apply: bool, with_sudo: bool, with_polkit: bool) -> ExitCode {
@@ -1850,884 +1503,6 @@ fn wire_service(
 
 // ---- pure PAM-text mechanics (unit-tested) -----------------------------------
 
-fn content_has_module(c: &str) -> bool {
-    c.lines().any(|l| {
-        let t = l.trim_start();
-        !t.starts_with('#') && t.contains(MODULE)
-    })
-}
-
-/// An `auth`-phase line whose password path is an `include` a `success=N` jump
-/// can't skip: Debian's `@include common-auth`/`login`, Arch's
-/// `auth include system-login`/`system-local-login`/`system-auth`, or a bare
-/// `auth include common-auth`. These need the `sufficient` (module IGNOREs on
-/// cold login) form, NOT the jump form. A `substack` is atomic for jump
-/// counting, so it deliberately does not match here and keeps the jump stanza —
-/// which is what openSUSE's `auth substack common-auth` relies on.
-fn is_include_auth_layout(line: &str) -> bool {
-    let t = line.trim_start();
-    if t.starts_with("@include common-auth") || t.starts_with("@include login") {
-        return true;
-    }
-    let toks: Vec<&str> = t.split_whitespace().collect();
-    toks.first() == Some(&"auth")
-        && toks.get(1) == Some(&"include")
-        && matches!(
-            toks.get(2),
-            Some(&"system-login")
-                | Some(&"system-local-login")
-                | Some(&"system-auth")
-                | Some(&"common-auth")
-        )
-}
-
-/// `<kind>` is `auth`/`session`; matches the shared password stack that the
-/// `success=1` jump skips: Fedora's `password-auth`/`system-auth`, and
-/// openSUSE's `common-auth`/`common-session`.
-///
-/// The stack names are kind-aware so an `auth` line is only tested against
-/// auth-phase names. openSUSE's `plasmalogin` routes the password through
-/// `auth substack common-auth`, which matched nothing here: wiring then fell
-/// back to the first auth line and inserted the jump above `pam_nologin.so`, so
-/// a face login skipped the nologin gate and *still* landed on the password
-/// stack underneath — face auth that neither honoured nologin nor logged you in.
-fn is_passwd_substack(line: &str, kind: &str) -> bool {
-    let t = line.trim_start();
-    if t.starts_with('#') {
-        return false;
-    }
-    let toks: Vec<&str> = t
-        .strip_prefix('-')
-        .unwrap_or(t)
-        .split_whitespace()
-        .collect();
-    let stacks: &[&str] = match kind {
-        "auth" => &["password-auth", "system-auth", "common-auth"],
-        "session" => &["password-auth", "system-auth", "common-session"],
-        _ => &["password-auth", "system-auth"],
-    };
-    toks.first() == Some(&kind)
-        && toks.iter().any(|w| *w == "substack" || *w == "include")
-        && toks.iter().any(|w| stacks.contains(w))
-}
-
-fn is_auth_directive(line: &str) -> bool {
-    let t = line.trim_start();
-    if t.starts_with('#') {
-        return false;
-    }
-    t.strip_prefix('-').unwrap_or(t).split_whitespace().next() == Some("auth")
-}
-
-/// An `auth` line whose control keyword is `substack`, whatever the shared stack
-/// happens to be NAMED. A substack is atomic for jump counting, so this is a
-/// safe jump anchor even when we do not recognize the target.
-///
-/// This exists because the named list cannot keep up with upstreams. GDM's main
-/// branch renamed its shared stack from `password-auth` to
-/// `gdm-password-auth-substack` (a file GDM does not ship — distros supply it),
-/// which no name in `is_passwd_substack` matches. Without this tier the anchor
-/// search falls through to "first auth line", which on GDM's stack is
-/// `pam_selinux_permit.so`: the jump would then skip THAT and land above the
-/// password substack, which still runs. That is the openSUSE failure exactly,
-/// and it would arrive silently with a GDM upgrade.
-fn is_auth_substack_anchor(line: &str) -> bool {
-    let t = line.trim_start();
-    if t.starts_with('#') {
-        return false;
-    }
-    let toks: Vec<&str> = t
-        .strip_prefix('-')
-        .unwrap_or(t)
-        .split_whitespace()
-        .collect();
-    toks.first() == Some(&"auth") && toks.get(1) == Some(&"substack")
-}
-
-/// Where the face block anchors, in descending order of confidence: a shared
-/// stack we recognize by name, then any `substack` whatever its name, and only
-/// then the first `auth` line. The last tier is a guess and is kept last
-/// deliberately — it is what produces a jump over an unrelated module.
-fn find_auth_anchor(lines: &[&str]) -> Option<usize> {
-    lines
-        .iter()
-        .position(|l| is_passwd_substack(l, "auth"))
-        .or_else(|| lines.iter().position(|l| is_auth_substack_anchor(l)))
-        .or_else(|| lines.iter().position(|l| is_auth_directive(l)))
-}
-
-/// Login-keyring modules that CONSUME the password an `unseal` line releases.
-/// KDE's kwallet-pam still installs `pam_kwallet5.so` under Plasma 6 (the
-/// running provider process is `ksecretd`, but the PAM module kept the 5 in its
-/// name); older Plasma and a few distros ship `pam_kwallet.so`. GNOME ships
-/// `pam_gnome_keyring.so`.
-const KEYRING_CONSUMERS: &[&str] = &["pam_kwallet5.so", "pam_kwallet.so", "pam_gnome_keyring.so"];
-
-/// What a wired greeter stack does with the password irlume releases.
-///
-/// irlume's job ends at setting `PAM_AUTHTOK`. Turning that token into an OPEN
-/// wallet is another module's work, and if that module is absent (or sits above
-/// our line, where it runs before the token exists) the login succeeds by face
-/// and the wallet stays locked — which surfaces to the user as KWallet
-/// prompting for its password anyway. Nothing checked for this, so the failure
-/// was indistinguishable from "wired ✓".
-struct KeyringHandoff {
-    /// A module carrying BOTH halves: an `auth` line below ours (so it observes
-    /// the token) and a `session` line of its own. That is a hand-off that
-    /// actually opens a wallet.
-    ///
-    /// The two halves are paired PER MODULE rather than counted separately,
-    /// because upstream stacks list several consumers side by side: Fedora's
-    /// `plasmalogin` ships `pam_gnome_keyring.so`, `pam_kwallet5.so` AND
-    /// `pam_kwallet.so`. Counting "some auth line" and "some session line"
-    /// independently would call a stack complete when one module read the token
-    /// and a different one started a daemon that never received it.
-    complete: Option<&'static str>,
-    /// Modules with an auth line below ours but no session line of their own.
-    /// `pam_kwallet.c` stashes the key with `pam_set_data` in
-    /// `pam_sm_authenticate` and only `pam_sm_open_session` acts on it, so this
-    /// half alone derives a key and then drops it.
-    auth_only: Vec<&'static str>,
-}
-
-/// Locate the keyring hand-off in a wired stack. `None` when this stack
-/// releases no credential at all (no `unseal` line), which is how the plain
-/// verify surfaces — sudo and polkit — opt out of being judged against a wallet
-/// they were never meant to open. The lock screen opts out a level up instead:
-/// `report_keyring_handoff` walks only `GREETERS`, because a warm screen unlock
-/// runs against a wallet the login already opened.
-fn keyring_handoff(content: &str) -> Option<KeyringHandoff> {
-    let lines: Vec<&str> = content.lines().collect();
-    let unseal_at = lines.iter().position(|l| {
-        let t = l.trim_start();
-        !t.starts_with('#') && t.contains(MODULE) && t.contains("unseal")
-    })?;
-    let consumer_in = |l: &str| -> Option<&'static str> {
-        let t = l.trim_start();
-        if t.starts_with('#') {
-            return None;
-        }
-        KEYRING_CONSUMERS.iter().copied().find(|m| t.contains(m))
-    };
-    // A given module's session line may sit anywhere in the session phase, so
-    // that half is searched across the whole file; only the AUTH half is
-    // order-sensitive.
-    let has_session_line = |module: &str| {
-        lines.iter().any(|l| {
-            let t = l.trim_start();
-            if t.starts_with('#') {
-                return false;
-            }
-            let phase = t.strip_prefix('-').unwrap_or(t);
-            phase.split_whitespace().next() == Some("session") && t.contains(module)
-        })
-    };
-    // Only lines BELOW ours can see the token we set, so the search starts past
-    // the unseal line rather than scanning the whole file.
-    let mut complete = None;
-    let mut auth_only: Vec<&'static str> = Vec::new();
-    for module in lines
-        .iter()
-        .skip(unseal_at + 1)
-        .filter(|l| is_auth_directive(l))
-        .filter_map(|l| consumer_in(l))
-    {
-        if has_session_line(module) {
-            complete = complete.or(Some(module));
-        } else if !auth_only.contains(&module) {
-            auth_only.push(module);
-        }
-    }
-    Some(KeyringHandoff {
-        complete,
-        auth_only,
-    })
-}
-
-/// Print a warning for every wired greeter that releases the login password
-/// with nothing downstream to open the wallet. Advisory only: it never changes
-/// a stack and never fails a command, because a missing wallet module is the
-/// user's package/desktop choice, not a broken irlume wiring.
-fn report_keyring_handoff() {
-    for s in GREETERS {
-        let Some(path) = service_present(s) else {
-            continue;
-        };
-        let Ok(content) = std::fs::read_to_string(&path) else {
-            continue;
-        };
-        if !content_has_module(&content) {
-            continue;
-        }
-        let Some(handoff) = keyring_handoff(&content) else {
-            continue;
-        };
-        // A single complete module is enough: the wallet opens. Only when none
-        // is complete does the stack need explaining.
-        if handoff.complete.is_some() {
-            continue;
-        }
-        match handoff.auth_only.first() {
-            Some(m) => println!(
-                "  ⚠ {}: {m} reads the released password but has no session line, so the\n     \
-                 wallet daemon is never started with the key. Your wallet will still prompt.",
-                s.etc
-            ),
-            None => println!(
-                "  ⚠ {}: face login releases your login password, but no keyring module\n     \
-                 reads it afterwards, so KWallet/the login keyring will still prompt.\n     \
-                 Install kwallet-pam (KDE) or gnome-keyring (GNOME); if it is already\n     \
-                 installed, its auth line must sit BELOW the pam_irlume unseal line.",
-                s.etc
-            ),
-        }
-    }
-}
-
-/// Insert irlume's greeter block: `unseal` before the password substack, a
-/// `pam_permit` landing + `reseal` after it, and a `session reseal` after the
-/// session substack. Idempotent; falls back to the first `auth` line if there's
-/// no password substack.
-/// Wire a display-manager greeter. `face` adds the face-first login lines
-/// (Secure-tier credential release); `keyring` adds the post-auth keyring-unseal
-/// line (fingerprint keyring unlock; needed in gdm-password too, since GDM's
-/// SESSION keyring unlock runs through gdm-password even on a fingerprint login).
-/// Reseal (self-heal of the sealed password) rides along whenever either is set.
-fn wire_greeter_impl(content: &str, face: bool, keyring: bool, ondemand: bool) -> (String, bool) {
-    if !face && !keyring {
-        return (content.to_string(), false);
-    }
-    if content_has_module(content) {
-        return (content.to_string(), false);
-    }
-    let lines: Vec<&str> = content.lines().collect();
-    // Debian/Ubuntu layout: face-first `sufficient` before the password path;
-    // keyring-unseal after it (runs on any auth success, incl. a fingerprint via
-    // common-auth's pam_fprintd). Most greeters `@include common-auth` directly;
-    // greetd instead `@include login` (which itself pulls in common-auth) and adds
-    // its own keyring modules after; inserting the face line before that include
-    // works identically (face IGNORE on cold login → the include's pam_unix +
-    // greetd's pam_gnome_keyring run with the unsealed AUTHTOK → keyring unlocks).
-    if let Some(inc_at) = lines.iter().position(|l| is_include_auth_layout(l)) {
-        let mut out = Vec::with_capacity(lines.len() + 4);
-        for (i, l) in lines.iter().enumerate() {
-            if i == inc_at {
-                if face {
-                    out.push(include_greeter_line(
-                        if ondemand { "ondemand" } else { "facefirst" },
-                        true,
-                    ));
-                }
-                out.push((*l).to_string());
-                if keyring {
-                    out.push(KEYRING_UNSEAL.to_string());
-                }
-                out.push(RESEAL_AUTH.to_string());
-            } else if l.trim_start().starts_with("@include common-session") {
-                out.push((*l).to_string());
-                out.push(RESEAL_SESSION.to_string());
-            } else {
-                out.push((*l).to_string());
-            }
-        }
-        if !out.iter().any(|l| l == RESEAL_SESSION) {
-            out.push(RESEAL_SESSION.to_string());
-        }
-        return (format!("{}\n", out.join("\n")), true);
-    }
-    let auth_at = find_auth_anchor(&lines);
-    let sess_at = lines.iter().position(|l| is_passwd_substack(l, "session"));
-    let Some(auth_at) = auth_at else {
-        return (content.to_string(), false);
-    };
-    let mut out = Vec::with_capacity(lines.len() + 5);
-    for (i, l) in lines.iter().enumerate() {
-        if i == auth_at {
-            if face {
-                out.push(
-                    if ondemand {
-                        GREETER_UNSEAL_COSMIC_JUMP
-                    } else {
-                        GREETER_UNSEAL_FACEFIRST_JUMP
-                    }
-                    .to_string(),
-                );
-                out.push((*l).to_string());
-                out.push(PERMIT_LANDING.to_string());
-            } else {
-                out.push((*l).to_string());
-            }
-            if keyring {
-                out.push(KEYRING_UNSEAL.to_string());
-            }
-            out.push(RESEAL_AUTH.to_string());
-        } else if Some(i) == sess_at {
-            out.push((*l).to_string());
-            out.push(RESEAL_SESSION.to_string());
-        } else {
-            out.push((*l).to_string());
-        }
-    }
-    if sess_at.is_none() {
-        out.push(RESEAL_SESSION.to_string()); // harmless optional session line
-    }
-    (format!("{}\n", out.join("\n")), true)
-}
-
-/// Wire the KDE lock (`kde`) with the consent-driven on-demand face block: face
-/// engages only on an empty-field Enter, verifies identity for the unlock, and
-/// otherwise falls through to the password. Same `ondemand` mode as
-/// cosmic-greeter, applied to KDE's submit-driven lock service. No reseal (a
-/// screen unlock releases no credential). Handles both the Debian `@include`
-/// and the Fedora `substack` layouts.
-fn wire_lock(content: &str) -> (String, bool) {
-    if content_has_module(content) {
-        return (content.to_string(), false);
-    }
-    let lines: Vec<&str> = content.lines().collect();
-    // Include layout (Debian `@include common-auth`, Arch `auth include
-    // system-local-login`) → face-first `sufficient` before it. A warm lock so
-    // no keyring-continue arg; on face success the module returns SUCCESS and
-    // `sufficient` grants the unlock.
-    if let Some(inc_at) = lines.iter().position(|l| is_include_auth_layout(l)) {
-        let mut out = Vec::with_capacity(lines.len() + 1);
-        for (i, l) in lines.iter().enumerate() {
-            if i == inc_at {
-                out.push(include_greeter_line("ondemand", false));
-            }
-            out.push((*l).to_string());
-        }
-        return (format!("{}\n", out.join("\n")), true);
-    }
-    // Fedora `substack password-auth` layout → jump stanza + permit landing.
-    let auth_at = find_auth_anchor(&lines);
-    let Some(auth_at) = auth_at else {
-        return (content.to_string(), false);
-    };
-    let mut out = Vec::with_capacity(lines.len() + 2);
-    for (i, l) in lines.iter().enumerate() {
-        if i == auth_at {
-            out.push(GREETER_UNSEAL_COSMIC_JUMP.to_string());
-            out.push((*l).to_string());
-            out.push(PERMIT_LANDING.to_string());
-        } else {
-            out.push((*l).to_string());
-        }
-    }
-    (format!("{}\n", out.join("\n")), true)
-}
-
-/// Wire the `keyring` unseal line into a fingerprint login service
-/// (`gdm-fingerprint`): insert it right after the `pam_fprintd.so` auth line so
-/// the sealed password is set before pam_gnome_keyring's auth line runs.
-fn wire_fp_keyring(content: &str) -> (String, bool) {
-    if content.lines().any(|l| {
-        let t = l.trim_start();
-        !t.starts_with('#') && t.contains("pam_irlume.so") && t.contains("keyring")
-    }) {
-        return (content.to_string(), false); // already wired
-    }
-    let lines: Vec<&str> = content.lines().collect();
-    let Some(fp_at) = lines.iter().position(|l| is_fingerprint_auth(l)) else {
-        return (content.to_string(), false);
-    };
-    // Does anything in this stack already read the token we are about to set?
-    // GDM's own `gdm-fingerprint` does not, so the unseal line alone would
-    // release a password into a stack with no consumer.
-    let has_consumer = lines.iter().any(|l| {
-        let t = l.trim_start();
-        !t.starts_with('#') && KEYRING_CONSUMERS.iter().any(|m| t.contains(m))
-    });
-    let mut out = Vec::with_capacity(lines.len() + 3);
-    for (i, l) in lines.iter().enumerate() {
-        out.push((*l).to_string());
-        if i == fp_at {
-            out.push(KEYRING_UNSEAL.to_string());
-            if !has_consumer {
-                out.push(FP_GKR_AUTH.to_string());
-            }
-        }
-    }
-    if !has_consumer {
-        // Appended last: gnome-keyring's session half wants to run once the
-        // session is otherwise set up, and it is the half that actually starts
-        // the daemon and unlocks with the stashed key.
-        out.push(FP_GKR_SESSION.to_string());
-    }
-    (format!("{}\n", out.join("\n")), true)
-}
-
-/// The auth line that performs the fingerprint check, and therefore the line the
-/// keyring unseal must follow.
-///
-/// Matching a literal `pam_fprintd.so` was not enough: GDM's shipped
-/// `gdm-fingerprint.pam` never names the module, delegating instead to
-/// `auth substack fingerprint-auth` (renamed `gdm-fingerprint-auth-substack` on
-/// GDM's development branch). With only the literal match this returned no
-/// anchor, `wire_fp_keyring` became a silent no-op, and the fingerprint keyring
-/// unlock never wired on Fedora at all.
-fn is_fingerprint_auth(line: &str) -> bool {
-    let t = line.trim_start();
-    if t.starts_with('#') {
-        return false;
-    }
-    let toks: Vec<&str> = t
-        .strip_prefix('-')
-        .unwrap_or(t)
-        .split_whitespace()
-        .collect();
-    if toks.first() != Some(&"auth") {
-        return false;
-    }
-    toks.iter().any(|w| w.contains("pam_fprintd.so"))
-        || (toks.iter().any(|w| *w == "substack" || *w == "include")
-            && toks.iter().any(|w| w.contains("fingerprint")))
-}
-/// Wire a single-stanza verify service (`sudo`, `polkit-1`): the stanza goes
-/// ABOVE the first auth-phase line, whether that is Fedora's `auth include
-/// system-auth` or Debian/Ubuntu's `@include common-auth`. An anchor that only
-/// matched a literal `auth` token missed the include layout entirely and the
-/// stanza got appended at EOF, i.e. AFTER the password modules, where it is dead:
-/// a wrong password already hit common-auth's pam_deny, a right one already
-/// granted via pam_unix. No anchor at all → no wiring: appending to a file with
-/// no auth phase would leave pam_irlume as the only auth module, and its IGNORE
-/// on a failed face would then fail the whole prompt instead of falling back to
-/// the password.
-fn wire_verify_service(content: &str) -> (String, bool) {
-    if content_has_module(content) {
-        return (content.to_string(), false);
-    }
-    let lines: Vec<&str> = content.lines().collect();
-    let anchor = lines
-        .iter()
-        .position(|l| is_include_auth_layout(l) || is_auth_directive(l));
-    let Some(anchor) = anchor else {
-        return (content.to_string(), false);
-    };
-    let mut out = Vec::with_capacity(lines.len() + 1);
-    for (i, l) in lines.iter().enumerate() {
-        if i == anchor {
-            out.push(VERIFY_STANZA.to_string());
-        }
-        out.push((*l).to_string());
-    }
-    (format!("{}\n", out.join("\n")), true)
-}
-
-/// Remove every irlume line AND the pam_permit landing we added (used only when
-/// no backup exists; the backup-restore path is preferred).
-fn unwire_lines(content: &str) -> (String, bool) {
-    // Strip every pam_irlume line, plus ONLY the pam_permit landing WE tagged
-    // (`# irlume-landing`), never a foreign pam_permit.so.
-    let mut changed = false;
-    let kept: Vec<&str> = content
-        .lines()
-        .filter(|l| {
-            let t = l.trim_start();
-            if t.starts_with('#') {
-                return true;
-            }
-            let drop = t.contains(MODULE)
-                || (t.contains("pam_permit.so") && l.contains("# irlume-landing"))
-                // Only the gnome-keyring lines WE tagged; a distro-shipped
-                // keyring line carries no tag and must survive unwiring.
-                || (t.contains("pam_gnome_keyring.so") && l.contains(KEYRING_TAG));
-            if drop {
-                changed = true;
-            }
-            !drop
-        })
-        .collect();
-    (format!("{}\n", kept.join("\n")), changed)
-}
-
-// ---- file ops ----------------------------------------------------------------
-
-fn read(p: &str) -> Result<String, String> {
-    std::fs::read_to_string(p).map_err(|e| format!("read {p}: {e}"))
-}
-fn file_has_module(p: &Path) -> bool {
-    std::fs::read_to_string(p)
-        .map(|c| content_has_module(&c))
-        .unwrap_or(false)
-}
-fn file_is_created_override(p: &Path) -> bool {
-    std::fs::read_to_string(p)
-        .map(|c| c.starts_with(CREATED_PREFIX))
-        .unwrap_or(false)
-}
-fn service_present(s: &Svc) -> Option<PathBuf> {
-    if Path::new(s.etc).exists() {
-        return Some(PathBuf::from(s.etc));
-    }
-    s.vendor
-        .filter(|v| Path::new(v).exists())
-        .map(|_| PathBuf::from(s.etc))
-}
-/// Held for as long as a process is changing PAM. Released when dropped.
-pub(crate) struct PamLock {
-    _file: std::fs::File,
-}
-
-/// Where the PAM lock lives. `IRLUME_PAM_LOCK` overrides it for tests and
-/// containers, the same way `IRLUME_STATE_DIR` overrides the state root.
-fn pam_lock_path() -> PathBuf {
-    std::env::var_os("IRLUME_PAM_LOCK")
-        .map(PathBuf::from)
-        .unwrap_or_else(|| PathBuf::from("/run/lock/irlume-pam.lock"))
-}
-
-/// Take the exclusive lock every irlume path that changes PAM must hold.
-///
-/// Nothing serialised these before. `login apply`, `login rollback`, human
-/// `login enable`/`disable`, and `reconcile` could all run at once, and the
-/// combinations are not theoretical: the reconcile path unit fires when a PAM
-/// file changes, which is exactly what the other three do. Two of them
-/// interleaving produced a stack that was a mixture of both, and the record
-/// written by either then described a machine state that never existed.
-///
-/// The lock covers the whole operation, not each write: revalidating a plan,
-/// writing the prepared record, every PAM and sidecar write, and the confirming
-/// record all have to be one indivisible unit, or the record still describes
-/// something other than what is on disk.
-///
-/// `flock` is released by the kernel when the process exits however it exits, so
-/// a killed irlume does not strand it. It does not exclude package managers or
-/// an administrator with an editor: only irlume takes it, which is why every
-/// path still re-checks the file it is about to write.
-pub(crate) fn lock_pam() -> Result<PamLock, String> {
-    use std::os::unix::io::AsRawFd as _;
-    let path = pam_lock_path();
-    if let Some(dir) = path.parent() {
-        std::fs::create_dir_all(dir).map_err(|e| format!("create {}: {e}", dir.display()))?;
-    }
-    let file = std::fs::OpenOptions::new()
-        .write(true)
-        .create(true)
-        .truncate(false)
-        .open(&path)
-        .map_err(|e| format!("open {}: {e}", path.display()))?;
-    let fd = file.as_raw_fd();
-    // SAFETY: `fd` is owned by `file`, which outlives the call and the guard.
-    let busy = unsafe { libc::flock(fd, libc::LOCK_EX | libc::LOCK_NB) } != 0;
-    if busy {
-        // Said on stderr, because machine output is JSON on stdout. Then wait:
-        // refusing outright would make the reconcile path unit give up exactly
-        // when an apply is in flight, which is when it most needs to run after.
-        eprintln!("irlume: another irlume PAM operation is in progress, waiting for it…");
-        // SAFETY: as above.
-        if unsafe { libc::flock(fd, libc::LOCK_EX) } != 0 {
-            return Err(format!(
-                "lock {}: {}",
-                path.display(),
-                std::io::Error::last_os_error()
-            ));
-        }
-    }
-    sweep_abandoned_scratch();
-    Ok(PamLock { _file: file })
-}
-
-/// Remove scratch files a killed irlume left in `/etc/pam.d`.
-///
-/// A `SIGKILL` between creating the scratch file and renaming it skips every
-/// cleanup path, so real hardware runs that interrupt an apply leave
-/// `.sudo.irlume-new.1234.0.tmp` behind. PAM selects a stack by exact filename,
-/// so a dotfile is never read as a service and this is litter rather than a
-/// hazard — but it is irlume's litter, and it accumulates.
-///
-/// Done while holding the lock, which is what makes it safe: the name is one
-/// only this module produces, and no other irlume can be mid-write. Nothing
-/// outside that pattern is ever considered, because a cleanup that reasons about
-/// what "looks unexpected" is how a harness in this project deleted a real
-/// conffile.
-fn sweep_abandoned_scratch() {
-    let dir = std::path::Path::new("/etc/pam.d");
-    let Ok(entries) = std::fs::read_dir(dir) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let Some(name) = name.to_str() else { continue };
-        if name.starts_with('.') && name.contains(".irlume-") && name.ends_with(".tmp") {
-            let _ = std::fs::remove_file(entry.path());
-        }
-    }
-}
-
-/// Test-only: replace a target during the window between the first look and the
-/// rename, so the recheck has something to catch.
-///
-/// The window cannot be reached from outside — it is entirely inside one
-/// function call — so a test that only sets up a symlink beforehand proves the
-/// FIRST check, never the second. Without this, removing the pre-rename recheck
-/// left every test green.
-#[cfg(test)]
-fn swap_target_for_test(path: &Path) {
-    let mut armed = SWAP_DURING_WRITE.lock().unwrap_or_else(|e| e.into_inner());
-    if armed.as_deref() != Some(path) {
-        return;
-    }
-    *armed = None;
-    // A different inode under the same name: what an administrator, a package
-    // or another writer does in that window.
-    //
-    // Written elsewhere and renamed over, NOT removed and recreated. Removing
-    // frees the inode number, and a filesystem is free to hand the same one
-    // straight back: this test passed locally and failed in CI for exactly that
-    // reason. Both files exist at once here, so the numbers cannot coincide.
-    let replacement = path.with_extension("irlume-swap-source");
-    let _ = std::fs::write(&replacement, "SOMEONE ELSE'S FILE\n");
-    let _ = std::fs::rename(&replacement, path);
-}
-
-#[cfg(test)]
-static SWAP_DURING_WRITE: std::sync::Mutex<Option<PathBuf>> = std::sync::Mutex::new(None);
-
-/// What a PAM path is, for deciding whether irlume may replace it.
-///
-/// `None` means it does not exist, which is a legitimate state: apply removes an
-/// override it created, and rollback recreates a file that was absent.
-type TargetState = Option<(u64, u64)>;
-
-/// Establish that a PAM path is something irlume may replace, and identify it.
-///
-/// Two things are refused, and previously each was refused in one place or in
-/// none:
-///
-/// - **A symlink.** Renaming over it REPLACES the link with a regular file, and
-///   a rollback restores content rather than the link, so the conversion is
-///   silent and permanent. Writing through it instead is no better: on Fedora
-///   these point into `/etc/authselect` and on Debian into `/etc/alternatives`,
-///   shared targets other tooling owns. `apply` checked this; human
-///   enable/disable, reconcile and rollback did not, so one command refused a
-///   file another would quietly convert.
-/// - **More than one link to the inode.** A rename replaces one directory entry;
-///   every other name for that inode keeps referring to the OLD content. PAM
-///   then reads one inode while package tooling updates another. The link
-///   topology is recorded nowhere, so irlume could not put it back, which makes
-///   breaking it silently the wrong default.
-///
-/// The identity returned is the device and inode, which is what the caller
-/// compares to decide the name still refers to the same file. That is the usual
-/// answer and not a perfect one: a filesystem may hand the same inode number
-/// back for a file created right after the old one was unlinked, so a
-/// replacement can in principle wear the identity of what it replaced. irlume's
-/// own paths cannot collide here because they hold the PAM lock; against an
-/// external writer this narrows the window rather than closing it.
-fn inspect_target(path: &Path) -> Result<TargetState, String> {
-    use std::os::unix::fs::MetadataExt as _;
-    let meta = match std::fs::symlink_metadata(path) {
-        Ok(meta) => meta,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(e) => return Err(format!("stat {}: {e}", path.display())),
-    };
-    if meta.file_type().is_symlink() {
-        return Err(format!(
-            "{} is a symlink; irlume will not replace it with a regular file, because that \
-             conversion cannot be undone and the target belongs to another tool \
-             (authselect, alternatives)",
-            path.display()
-        ));
-    }
-    if !meta.file_type().is_file() {
-        return Err(format!("{} is not a regular file", path.display()));
-    }
-    if meta.nlink() > 1 {
-        return Err(format!(
-            "{} has {} hard links; replacing it would leave the other names referring to the \
-             old content, and irlume does not record the link topology so it could not put \
-             it back",
-            path.display(),
-            meta.nlink()
-        ));
-    }
-    Ok(Some((meta.dev(), meta.ino())))
-}
-
-/// A scratch path in the same directory as `path`, unique to this call.
-///
-/// Every write here used to share one name per service, `.{service}.irlume.tmp`.
-/// Two irlume processes writing the same PAM file would open that one inode and
-/// interleave their bodies, and whichever renamed first published whatever was
-/// in it: an atomic rename makes the NAME change indivisible, it does not make
-/// concurrent production of the source safe. The PAM lock now keeps irlume's own
-/// paths apart, and a unique name means a leftover from a killed run is never
-/// adopted either.
-fn scratch_path(path: &Path, kind: &str) -> PathBuf {
-    static SEQ: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
-    let seq = SEQ.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    let dir = path.parent().unwrap_or_else(|| Path::new("."));
-    let fname = path.file_name().and_then(|s| s.to_str()).unwrap_or("pam");
-    dir.join(format!(
-        ".{fname}.irlume-{kind}.{}.{seq}.tmp",
-        std::process::id()
-    ))
-}
-
-/// Create the scratch file, never adopting one that is already there.
-fn create_scratch(tmp: &Path) -> Result<std::fs::File, String> {
-    let open = || {
-        std::fs::OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .open(tmp)
-    };
-    match open() {
-        // Same pid and counter as a crashed earlier run. Drop it rather than
-        // write into a file whose contents are somebody else's.
-        Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => {
-            std::fs::remove_file(tmp).map_err(|e| format!("remove {}: {e}", tmp.display()))?;
-            open().map_err(|e| format!("create {}: {e}", tmp.display()))
-        }
-        other => other.map_err(|e| format!("create {}: {e}", tmp.display())),
-    }
-}
-
-/// Make a directory durable, so an entry created in it survives a power loss.
-fn fsync_dir(dir: &Path) -> Result<(), String> {
-    std::fs::File::open(dir)
-        .and_then(|d| d.sync_all())
-        .map_err(|e| format!("fsync {}: {e}", dir.display()))
-}
-
-/// Copy `path` to its `.pre-irlume` backup, atomically, if there is not one yet.
-///
-/// The copy used to go straight to the final name. A kill or an ENOSPC part way
-/// through left a TRUNCATED file at `.pre-irlume`, and the next enable treats an
-/// existing backup as the pristine origin to rebuild from, so a half-copied
-/// stack became the authority for what the machine's PAM should contain. A
-/// backup that only ever appears complete cannot be believed part way.
-///
-/// The destination is published with `hard_link`, which fails if the name
-/// already exists rather than replacing it. An `exists()` test followed by a
-/// rename would be the same check-then-act split this file has been bitten by
-/// before, and would let a retry overwrite a good backup with the already-wired
-/// content.
-fn backup(path: &Path) -> Result<(), String> {
-    let bak = PathBuf::from(format!("{}{BACKUP}", path.display()));
-    // The backup is held to the same standard as the stack it came from, and it
-    // was not. `exists()` follows a symlink, so a `.pre-irlume` pointing
-    // somewhere else was accepted and then used as the pristine origin a later
-    // enable rebuilds from. A DANGLING one was worse: `exists()` said no, and the
-    // publishing link then failed with EEXIST against the symlink's own name,
-    // which read as "a backup is already there" when there was none at all.
-    // A complete backup already there is left alone; it must not be replaced
-    // with the now-wired content.
-    if inspect_target(&bak)?.is_some() {
-        return Ok(());
-    }
-    let contents = std::fs::read(path).map_err(|e| format!("read {}: {e}", path.display()))?;
-    let meta =
-        std::fs::symlink_metadata(path).map_err(|e| format!("stat {}: {e}", path.display()))?;
-    let tmp = scratch_path(path, "bak");
-    let written = (|| -> Result<(), String> {
-        use std::io::Write as _;
-        let mut file = create_scratch(&tmp)?;
-        file.write_all(&contents)
-            .map_err(|e| format!("write {}: {e}", tmp.display()))?;
-        apply_metadata(&tmp, &meta)?;
-        // Before the link, so the name never points at bytes that are not there.
-        file.sync_all()
-            .map_err(|e| format!("fsync {}: {e}", tmp.display()))?;
-        // Fails with EEXIST if another run got there first, which is the answer
-        // wanted: that backup is complete and this one is redundant.
-        match std::fs::hard_link(&tmp, &bak) {
-            Ok(()) => {}
-            Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => return Ok(()),
-            Err(e) => return Err(format!("backup {}: {e}", path.display())),
-        }
-        fsync_dir(path.parent().unwrap_or_else(|| Path::new(".")))
-    })();
-    let _ = std::fs::remove_file(&tmp);
-    written
-}
-
-/// Copy mode and ownership onto a path.
-fn apply_metadata(path: &Path, meta: &std::fs::Metadata) -> Result<(), String> {
-    use std::os::unix::fs::MetadataExt as _;
-    use std::os::unix::fs::PermissionsExt as _;
-    std::fs::set_permissions(path, std::fs::Permissions::from_mode(meta.mode() & 0o7777))
-        .map_err(|e| format!("chmod {}: {e}", path.display()))?;
-    std::os::unix::fs::chown(path, Some(meta.uid()), Some(meta.gid()))
-        .map_err(|e| format!("chown {}: {e}", path.display()))
-}
-
-fn write_atomic(path: &Path, contents: &str) -> Result<(), String> {
-    let existing = std::fs::symlink_metadata(path).ok();
-    write_atomic_inner(path, contents, existing.as_ref().map(mode_uid_gid))
-}
-
-fn mode_uid_gid(meta: &std::fs::Metadata) -> (u32, u32, u32) {
-    use std::os::unix::fs::MetadataExt as _;
-    (meta.mode() & 0o7777, meta.uid(), meta.gid())
-}
-
-/// Replace `path` with `contents`, durably, carrying the given mode and owner.
-///
-/// Attributes are set on the scratch file BEFORE the rename, so the name never
-/// resolves to a PAM file with the wrong mode or owner. Setting them afterwards
-/// leaves a window in which the live stack is whatever the root process's umask
-/// produced, and on the restore path the file did not exist to copy from at all.
-///
-/// `sync_all` before the rename and an fsync of `/etc/pam.d` after it: without
-/// them a successful `close` says nothing about what survives a power loss, and
-/// a PAM stack that comes back as a mixture of two versions is the failure this
-/// whole module exists to avoid.
-pub(crate) fn write_atomic_inner(
-    path: &Path,
-    contents: &str,
-    attrs: Option<(u32, u32, u32)>,
-) -> Result<(), String> {
-    use std::io::Write as _;
-    use std::os::unix::fs::PermissionsExt as _;
-    let dir = path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .to_path_buf();
-    // What the target is right now. A rename REPLACES whatever the name refers
-    // to, so this has to be settled before anything is written, and confirmed
-    // again before the name is taken over.
-    let before = inspect_target(path)?;
-    let tmp = scratch_path(path, "new");
-    let result = (|| -> Result<(), String> {
-        let mut file = create_scratch(&tmp)?;
-        file.write_all(contents.as_bytes())
-            .map_err(|e| format!("write {}: {e}", tmp.display()))?;
-        if let Some((mode, uid, gid)) = attrs {
-            std::fs::set_permissions(&tmp, std::fs::Permissions::from_mode(mode))
-                .map_err(|e| format!("chmod {}: {e}", tmp.display()))?;
-            // Ownership needs privilege, which every caller that writes PAM has,
-            // but an unusual filesystem can still refuse. A PAM file with the
-            // wrong group is a real access change, so it is reported.
-            std::os::unix::fs::chown(&tmp, Some(uid), Some(gid))
-                .map_err(|e| format!("chown {}: {e}", tmp.display()))?;
-        }
-        file.sync_all()
-            .map_err(|e| format!("fsync {}: {e}", tmp.display()))?;
-        drop(file);
-        #[cfg(test)]
-        swap_target_for_test(path);
-        // Immediately before the name is taken over, not once at the start. The
-        // first look and the rename are two moments, and what matters is what
-        // the name refers to at the instant it is replaced.
-        if inspect_target(path)? != before {
-            return Err(format!(
-                "{} changed while irlume was writing it, so it was left alone",
-                path.display()
-            ));
-        }
-        std::fs::rename(&tmp, path).map_err(|e| format!("rename into {}: {e}", path.display()))?;
-        fsync_dir(&dir)
-    })();
-    if result.is_err() {
-        let _ = std::fs::remove_file(&tmp);
-    }
-    result
-}
-
-// ---- SELinux (Fedora) --------------------------------------------------------
-
 /// `Some(true/false)` when semodule could be queried (root), `None` otherwise.
 fn selinux_loaded() -> Option<bool> {
     let out = Command::new("semodule").arg("-l").output().ok()?;
@@ -2832,6 +1607,9 @@ fn effective_uid() -> u32 {
 #[cfg(test)]
 mod tests {
     use super::*;
+    // Reached through the submodule because the parent has no production use
+    // for it; `use super::*` only carries what the parent itself imports.
+    use super::report::label_of;
 
     // Fedora gdm-password layout (real /etc file, the GDM greeter).
     const GDM: &str = "#%PAM-1.0\nauth     [success=done ...] pam_selinux_permit.so\nauth     substack      password-auth\nauth     optional      pam_gnome_keyring.so\naccount  include       password-auth\nsession  include       password-auth\nsession  optional      pam_gnome_keyring.so auto_start\n";
